Create main_code
Make sure to run this code on Arduino Uno on a TARC rocket. Adjust the parameters (target altitude, time, etc.) as necessary. Please note that we used a Teensy instead of an Arduino to decrease mass, but if you want to use an Arduino Uno instead, change all instances of Serial1 to Serial.